fix(volunteer-activities): 레이아웃, 스타일, 구조 개편
