+++ src/components/features/volunteer-activities/VolunteerActivityFilterTab.tsx
@@ -2,8 +2,6 @@
 
 import React, { useEffect, useState } from 'react'
 import {
-	Badge,
-	Button,
 	Input,
 	Select,
 	SelectContent,
@@ -11,7 +9,7 @@ import {
 	SelectTrigger,
 	SelectValue,
 } from '@/components/ui'
-import { Calendar, Clock, Filter, Search, X } from 'lucide-react'
+import { Filter, Search, X } from 'lucide-react'
 import {
 	VOLUNTEER_ACTIVITY_FILTER_OPTIONS,
 	VolunteerActivityFilterTabProps,
@@ -20,7 +18,6 @@ import {
 export default function ActivityFilterTab({
 	selectedStatus,
 	statusChangeAction,
-	allStatusCount,
 	searchQuery,
 	searchChangeAction,
 }: VolunteerActivityFilterTabProps) {
@@ -45,44 +42,12 @@ export default function ActivityFilterTab({
 		searchChangeAction('')
 	}
 
-	const handleClearAllFilters = () => {
-		statusChangeAction('all')
-		handleClearSearch()
-	}
-
-	const hasActiveFilters = selectedStatus !== 'all' || searchQuery.trim() !== ''
-	const selectedFilter = VOLUNTEER_ACTIVITY_FILTER_OPTIONS.find(
-		(opt) => opt.value === selectedStatus,
-	)
-
 	return (
-		<div className="space-y-4">
-			{/* 헤더 섹션 */}
-			<div className="flex items-center justify-between">
-				<div className="flex items-center gap-2">
-					<Calendar className="w-5 h-5 text-emerald-600" />
-					<h3 className="text-lg font-semibold text-gray-900 dark:text-gray-100">
-						활동 검색
-					</h3>
-				</div>
-
-				{hasActiveFilters && (
-					<Button
-						variant="ghost"
-						size="sm"
-						onClick={handleClearAllFilters}
-						className="text-gray-500 hover:text-gray-700 dark:text-gray-400 dark:hover:text-gray-200 h-8 px-3"
-					>
-						<X className="w-4 h-4 mr-1" />
-						초기화
-					</Button>
-				)}
-			</div>
-
+		<div>
 			{/* 검색 및 필터 섹션 */}
-			<div className="flex flex-col sm:flex-row gap-3">
+			<div className="flex flex-col sm:flex-row-reverse sm:justify-between sm:items-center gap-3">
 				{/* 검색바 */}
-				<div className="relative flex-1">
+				<div className="relative w-full sm:w-80">
 					<Search className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 w-4 h-4" />
 					<Input
 						type="text"
@@ -102,88 +67,37 @@ export default function ActivityFilterTab({
 					)}
 				</div>
 
-				{/* 상태 필터 드롭다운 */}
-				<div className="w-full sm:w-48">
-					<Select value={selectedStatus} onValueChange={statusChangeAction}>
-						<SelectTrigger className="h-10 bg-white dark:bg-gray-800 border-gray-200 dark:border-gray-700 rounded-lg">
-							<div className="flex items-center gap-2">
-								<Filter className="w-4 h-4 text-gray-500" />
-								<SelectValue placeholder="상태 선택" />
-							</div>
-						</SelectTrigger>
-						<SelectContent>
-							{VOLUNTEER_ACTIVITY_FILTER_OPTIONS.map((option) => {
-								const IconComponent = option.icon
-								return (
-									<SelectItem key={option.value} value={option.value}>
-										<div className="flex items-center gap-2">
-											<div className={`w-2 h-2 rounded-full ${option.color}`} />
-											<IconComponent className="w-4 h-4" />
-											<span>{option.label}</span>
-										</div>
-									</SelectItem>
-								)
-							})}
-						</SelectContent>
-					</Select>
+				{/* 오른쪽 필터 영역 */}
+				<div className="flex items-center gap-3">
+					{/* 상태 필터 드롭다운 */}
+					<div className="w-full sm:w-80">
+						<Select value={selectedStatus} onValueChange={statusChangeAction}>
+							<SelectTrigger className="h-10 bg-white dark:bg-gray-800 border-gray-200 dark:border-gray-700 rounded-lg">
+								<div className="flex items-center gap-2">
+									<Filter className="w-4 h-4 text-gray-500" />
+									<SelectValue placeholder="상태 선택" />
+								</div>
+							</SelectTrigger>
+							<SelectContent>
+								{VOLUNTEER_ACTIVITY_FILTER_OPTIONS.map((option) => {
+									const IconComponent = option.icon
+									return (
+										<SelectItem key={option.value} value={option.value}>
+											<div className="flex items-center gap-2">
+												<div
+													className={`w-2 h-2 rounded-full ${option.color}`}
+												/>
+												<IconComponent className="w-4 h-4" />
+												<span>{option.label}</span>
+											</div>
+										</SelectItem>
+									)
+								})}
+							</SelectContent>
+						</Select>
+					</div>
 				</div>
 			</div>
-
-			{/* 활성 필터 및 결과 표시 */}
-			<div className="flex flex-wrap items-center gap-2">
-				{/* 활성 필터 배지들 */}
-				{searchQuery.trim() && (
-					<Badge
-						variant="secondary"
-						className="bg-emerald-100 text-emerald-800 dark:bg-emerald-900/20 dark:text-emerald-300"
-					>
-						<Search className="w-3 h-3 mr-1" />
-						검색: {searchQuery}
-						<button
-							onClick={handleClearSearch}
-							className="ml-1 hover:bg-emerald-200 dark:hover:bg-emerald-800 rounded-full p-0.5"
-						>
-							<X className="w-3 h-3" />
-						</button>
-					</Badge>
-				)}
-
-				{selectedStatus !== 'all' && selectedFilter && (
-					<Badge
-						variant="secondary"
-						className="bg-blue-100 text-blue-800 dark:bg-blue-900/20 dark:text-blue-300"
-					>
-						<div
-							className={`w-2 h-2 rounded-full ${selectedFilter.color} mr-1`}
-						/>
-						상태: {selectedFilter.label}
-						<button
-							onClick={() => statusChangeAction('all')}
-							className="ml-1 hover:bg-blue-200 dark:hover:bg-blue-800 rounded-full p-0.5"
-						>
-							<X className="w-3 h-3" />
-						</button>
-					</Badge>
-				)}
-
-				{/* 결과 카운트 */}
-				<div className="flex items-center gap-1 text-sm text-gray-600 dark:text-gray-400 ml-auto">
-					<Clock className="w-4 h-4" />
-					<span className="font-medium">{allStatusCount}개</span>
-					<span>결과</span>
-				</div>
-			</div>
-
-			{/* 필터 설명 (필터가 적용된 경우만) */}
-			{hasActiveFilters && (
-				<div className="text-xs text-gray-500 dark:text-gray-400 bg-gray-50 dark:bg-gray-800/50 rounded-lg px-3 py-2">
-					{searchQuery.trim() && selectedStatus !== 'all'
-						? `"${searchQuery}" 검색 결과 중 ${selectedFilter?.label} 상태의 활동을 표시하고 있습니다.`
-						: searchQuery.trim()
-							? `"${searchQuery}" 검색 결과를 표시하고 있습니다.`
-							: `${selectedFilter?.label} 상태의 활동을 표시하고 있습니다.`}
-				</div>
-			)}
 		</div>
 	)
 }
